More work on the new Wavefront parser.

+++ src/utils/loaders/mtlloader.h
@@ -8,4 +8,4 @@
 
 #pragma once
 
-struct material *parseMTLFile(char *filePath, size_t *mtlCount);
+struct material *parseMTLFile(char *filePath, int *mtlCount);
